Consolidate STJ polymorphism tests into shared abstract suite
Moves PolymorphicTests.CustomTypeHierarchies.cs and PolymorphicTests.TypeClassifier.cs from tests/System.Text.Json.Tests/Serialization/ into tests/Common/ so they run under both reflection (11 wrappers) and source-gen (4 wrappers) via the SerializerWrapper pattern. Removes the duplicate PolymorphismTests.cs in the source-gen test project and wires the consolidated suite through a new tests/System.Text.Json.SourceGeneration.Tests/Serialization/PolymorphicTests.cs.

Test-only refactoring in preparation for #129041.
