+++ src/libraries/System.Text.Json/tests/System.Text.Json.SourceGeneration.Tests/Serialization/PolymorphicTests.cs
@@ -1,0 +1,493 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+using System.Text.Json.Nodes;
+using System.Text.Json.Serialization;
+using System.Text.Json.Serialization.Tests;
+
+namespace System.Text.Json.SourceGeneration.Tests
+{
+    public sealed partial class PolymorphicTests_Metadata : PolymorphicTests
+    {
+        public PolymorphicTests_Metadata()
+            : base(new StringSerializerWrapper(PolymorphicTestsContext_Metadata.Default))
+        {
+        }
+
+        [JsonSourceGenerationOptions(GenerationMode = JsonSourceGenerationMode.Metadata)]
+        [JsonSerializable(typeof(AttrClassifiedAnimal))]
+        [JsonSerializable(typeof(BinaryTree))]
+        [JsonSerializable(typeof(ClassifiedAnimalBase))]
+        [JsonSerializable(typeof(ClassifiedAnimalBase[]))]
+        [JsonSerializable(typeof(ClassifiedAnimalDeepHierarchy))]
+        [JsonSerializable(typeof(ClassifiedShape))]
+        [JsonSerializable(typeof(Dictionary<string, ClassifiedAnimalBase>))]
+        [JsonSerializable(typeof(DrawingCanvas))]
+        [JsonSerializable(typeof(FactoryClassifiedAnimal))]
+        [JsonSerializable(typeof(IEnumerable<int>[]))]
+        [JsonSerializable(typeof(IEnumerable<KeyValuePair<int, object>>[]))]
+        [JsonSerializable(typeof(IOpenGenericBase<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_DiamondA<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_DiamondB<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_InterfaceHierarchy<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_InterfaceWrapped<List<string>>))]
+        [JsonSerializable(typeof(IOpenGenericBase_MultiBaseA<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_MultiBaseB<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_MultiCtor<string>))]
+        [JsonSerializable(typeof(IVarCovBase<VarAnimal>))]
+        [JsonSerializable(typeof(List<ClassifiedAnimalBase>))]
+        [JsonSerializable(typeof(List<PolymorphicClass>))]
+        [JsonSerializable(typeof(List<PolymorphicClassWithCustomTypeDiscriminator>))]
+        [JsonSerializable(typeof(OpenGenericBase<string, int>))]
+        [JsonSerializable(typeof(OpenGenericBase_ArrayArg<int[]>))]
+        [JsonSerializable(typeof(OpenGenericBase_IntDisc<string>))]
+        [JsonSerializable(typeof(OpenGenericBase_KvpArg<KeyValuePair<string, int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_Mixed<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_Multi<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_MultiLevel<List<int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_Partial<string, int>))]
+        [JsonSerializable(typeof(OpenGenericBase_PartiallyConcrete<string, int>))]
+        [JsonSerializable(typeof(OpenGenericBase_Reordered<int, string>))]
+        [JsonSerializable(typeof(OpenGenericBase_StringDisc<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_Tuple<(int, string)>))]
+        [JsonSerializable(typeof(OpenGenericBase_Wrapped<List<string>>))]
+        [JsonSerializable(typeof(Peano))]
+        [JsonSerializable(typeof(PetOwner))]
+        [JsonSerializable(typeof(PolymorphicAbstractClass))]
+        [JsonSerializable(typeof(PolymorphicBaseWithCustomDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass))]
+        [JsonSerializable(typeof(PolymorphicClass[]))]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_NoTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_TypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass_InvalidCustomTypeDiscriminatorPropertyName))]
+        [JsonSerializable(typeof(PolymorphicClass_NoTypeDiscriminators))]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator[]))]
+        [JsonSerializable(typeof(PolymorphicClassWithDerivedCollections))]
+        [JsonSerializable(typeof(PolymorphicClassWithEscapedTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicDictionary))]
+        [JsonSerializable(typeof(PolymorphicInterface))]
+        [JsonSerializable(typeof(PolymorphicInterfaceWithInterfaceDerivedType))]
+        [JsonSerializable(typeof(PolymorphicIntList))]
+        [JsonSerializable(typeof(PolymorphicList))]
+        [JsonSerializable(typeof(PolymorphicTypeWithExtensionDataConflictingPropertyName))]
+        [JsonSerializable(typeof(PolymorphicTypeWithIgnoredConflictingPropertyName))]
+        [JsonSerializable(typeof(TestNode))]
+        [JsonSerializable(typeof(UnsupportedAttrClassifiedAnimal))]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClassWithCustomTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicInterfaceWithInterfaceDerivedType.DerivedClass), TypeInfoPropertyName = "PolymorphicInterfaceWithInterfaceDerivedType_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_CustomConverter_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_CustomConverter_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_InvalidCustomTypeDiscriminatorPropertyName.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_InvalidCustomTypeDiscriminatorPropertyName_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicAbstractClass.Derived), TypeInfoPropertyName = "PolymorphicAbstractClass_Derived")]
+        [JsonSerializable(typeof(PolymorphicClassWithEscapedTypeDiscriminator.Derived), TypeInfoPropertyName = "PolymorphicClassWithEscapedTypeDiscriminator_Derived")]
+        [JsonSerializable(typeof(PolymorphicTypeWithIgnoredConflictingPropertyName.Derived), TypeInfoPropertyName = "PolymorphicTypeWithIgnoredConflictingPropertyName_Derived")]
+        [JsonSerializable(typeof(PolymorphicTypeWithExtensionDataConflictingPropertyName.Derived), TypeInfoPropertyName = "PolymorphicTypeWithExtensionDataConflictingPropertyName_Derived")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass1_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass1_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass1_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass1_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass2_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass2_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass2_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass2_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedAbstractClass.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedAbstractClass_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedCollection_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedCollection_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedCollection_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedCollection_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedDictionary_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedDictionary_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedDictionary_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedDictionary_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClassWithConstructor_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClassWithConstructor_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClassWithCustomConverter_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClassWithCustomConverter_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClassWithCustomConverter_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClassWithCustomConverter_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicInterface.DerivedClass_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicInterface_DerivedClass_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicInterface.DerivedClass_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicInterface_DerivedClass_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(List<PolymorphicClassWithConstructor>))]
+        [JsonSerializable(typeof(List<PolymorphicDictionary>))]
+        [JsonSerializable(typeof(List<PolymorphicInterface>))]
+        [JsonSerializable(typeof(List<PolymorphicList>))]
+        [JsonSerializable(typeof(object))]
+        [JsonSerializable(typeof(Customer))]
+        [JsonSerializable(typeof(ObjectWithObjectProperties))]
+        [JsonSerializable(typeof(Person))]
+        [JsonSerializable(typeof(ConstraintBase<List<string>>))]
+        [JsonSerializable(typeof(IVarBivariantBase<VarDog, VarAnimal>))]
+        [JsonSerializable(typeof(IVarContraBase<VarDog>))]
+        [JsonSerializable(typeof(NestedBaseB<NestedOuterB<string>.NestedBoxB<int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_ComplexArg<List<int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_Nested<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_NullableArg<string>))]
+        [JsonSerializable(typeof(OpenGenericBase_StringDisc<string>))]
+        [JsonSerializable(typeof(PolymorphicAbstractClassWithAbstractClassDerivedType))]
+        [JsonSerializable(typeof(PolymorphicClasWithDuplicateTypeDiscriminators))]
+        [JsonSerializable(typeof(PolymorphicClassWithConstructor))]
+        [JsonSerializable(typeof(PolymorphicClassWithDuplicateDerivedTypeRegistrations))]
+        [JsonSerializable(typeof(PolymorphicClassWithNonAssignableDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClassWithObjectDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClassWithStructDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClassWithoutDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClass_WithBaseTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass_WithBaseTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_WithBaseTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_WithDerivedPolymorphicClass))]
+        [JsonSerializable(typeof(PolymorphicGenericClass<int>))]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtBase))]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtDerived))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicClass>>))]
+        [JsonSerializable(typeof(GenericPoco<Peano>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClass>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicDictionary>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicInterface>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicList>))]
+        [JsonSerializable(typeof(SimpleTestClassWithObject))]
+        [JsonSerializable(typeof(PolymorphicAbstractClassWithAbstractClassDerivedType.DerivedClass), TypeInfoPropertyName = "PolymorphicAbstractClassWithAbstractClassDerivedType_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClassWithConstructor.DerivedClass), TypeInfoPropertyName = "PolymorphicClassWithConstructor_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClassWithConstructor.DerivedCollection), TypeInfoPropertyName = "PolymorphicClassWithConstructor_DerivedCollection")]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator.DerivedCollection), TypeInfoPropertyName = "PolymorphicClassWithCustomTypeDiscriminator_DerivedCollection")]
+        [JsonSerializable(typeof(PolymorphicClassWithDuplicateDerivedTypeRegistrations.DerivedClass), TypeInfoPropertyName = "PolymorphicClassWithDuplicateDerivedTypeRegistrations_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_NoTypeDiscriminators.DerivedClass2), TypeInfoPropertyName = "PolymorphicClass_NoTypeDiscriminators_DerivedClass2")]
+        [JsonSerializable(typeof(PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_WithDerivedPolymorphicClass_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass.DerivedClass2), TypeInfoPropertyName = "PolymorphicClass_WithDerivedPolymorphicClass_DerivedClass_DerivedClass2")]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtBase.Derived), TypeInfoPropertyName = "PolymorphicTypeWithConflictingPropertyNameAtBase_Derived")]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtDerived.Derived), TypeInfoPropertyName = "PolymorphicTypeWithConflictingPropertyNameAtDerived_Derived")]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClass>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicDictionary>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicInterface>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicList>))]
+        [JsonSerializable(typeof(List<List<PolymorphicClass>>))]
+        [JsonSerializable(typeof(List<Peano>))]
+        [JsonSerializable(typeof(short[]))]
+        [JsonSerializable(typeof(object[]))]
+        [JsonSerializable(typeof(string[]))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicClassWithConstructor>>))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicDictionary>>))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicInterface>>))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicList>>))]
+        [JsonSerializable(typeof(GenericPoco<object>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClass_WithDerivedPolymorphicClass>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClassWithConstructor>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicClass>>))]
+        [JsonSerializable(typeof(Dictionary<string, string>))]
+        [JsonSerializable(typeof(Dictionary<string, Peano>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClassWithConstructor>))]
+        [JsonSerializable(typeof(HashSet<string>))]
+        [JsonSerializable(typeof(List<List<PolymorphicClassWithConstructor>>))]
+        [JsonSerializable(typeof(List<List<PolymorphicDictionary>>))]
+        [JsonSerializable(typeof(List<List<PolymorphicInterface>>))]
+        [JsonSerializable(typeof(List<List<PolymorphicList>>))]
+        [JsonSerializable(typeof(List<object>))]
+        [JsonSerializable(typeof(List<PolymorphicClass_WithDerivedPolymorphicClass>))]
+        [JsonSerializable(typeof(long[]))]
+        [JsonSerializable(typeof(JsonNode))]
+        [JsonSerializable(typeof(JsonObject))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicClassWithConstructor>>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicDictionary>>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicInterface>>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicList>>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClass_WithDerivedPolymorphicClass>))]
+        [JsonSerializable(typeof(Stack<string>))]
+        [JsonSerializable(typeof(System.Collections.IEnumerable))]
+        [JsonSerializable(typeof(ushort[]))]
+        [JsonSerializable(typeof(int[]))]
+        [JsonSerializable(typeof(uint[]))]
+        [JsonSerializable(typeof(ulong[]))]
+        [JsonSerializable(typeof(byte[]))]
+        [JsonSerializable(typeof(sbyte[]))]
+        [JsonSerializable(typeof(char[]))]
+        [JsonSerializable(typeof(bool[]))]
+        [JsonSerializable(typeof(float[]))]
+        [JsonSerializable(typeof(double[]))]
+        [JsonSerializable(typeof(decimal[]))]
+        [JsonSerializable(typeof(System.DateTime[]))]
+        [JsonSerializable(typeof(SampleEnum[]))]
+        [JsonSerializable(typeof(int?[]))]
+        [JsonSerializable(typeof(System.Guid))]
+        [JsonSerializable(typeof(SimpleStruct))]
+        [JsonSerializable(typeof(List<string>))]
+        [JsonSerializable(typeof(Queue<string>))]
+        [JsonSerializable(typeof(LinkedList<string>))]
+        [JsonSerializable(typeof(SortedSet<string>))]
+        [JsonSerializable(typeof(KeyValuePair<string, string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableArray<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableList<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableStack<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableQueue<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableHashSet<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableSortedSet<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableDictionary<string, string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableSortedDictionary<string, string>))]
+        [JsonSerializable(typeof(System.Collections.IList))]
+        [JsonSerializable(typeof(System.Collections.ICollection))]
+        [JsonSerializable(typeof(IEnumerable<object>))]
+        [JsonSerializable(typeof(IList<object>))]
+        [JsonSerializable(typeof(ICollection<object>))]
+        [JsonSerializable(typeof(IReadOnlyCollection<object>))]
+        [JsonSerializable(typeof(IReadOnlyList<object>))]
+        [JsonSerializable(typeof(ISet<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableList<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableStack<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableQueue<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableSet<object>))]
+        [JsonSerializable(typeof(HashSet<object>))]
+        [JsonSerializable(typeof(Stack<object>))]
+        [JsonSerializable(typeof(Queue<object>))]
+        [JsonSerializable(typeof(LinkedList<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableArray<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableList<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableStack<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableQueue<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableHashSet<object>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass>))]
+        [JsonSerializable(typeof(byte))]
+        [JsonSerializable(typeof(List<PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass>))]
+        [JsonSerializable(typeof(OpenGenericBase_DuplicateDerivedRegistrations<int>))]
+    internal sealed partial class PolymorphicTestsContext_Metadata : JsonSerializerContext
+        {
+        }
+    }
+
+    public sealed partial class PolymorphicTests_Metadata_AsyncStream : PolymorphicTests
+    {
+        public PolymorphicTests_Metadata_AsyncStream()
+            : base(new AsyncStreamSerializerWrapper(PolymorphicTests_Metadata.PolymorphicTestsContext_Metadata.Default))
+        {
+        }
+    }
+
+    public sealed partial class PolymorphicTests_Default : PolymorphicTests
+    {
+        public PolymorphicTests_Default()
+            : base(new StringSerializerWrapper(PolymorphicTestsContext_Default.Default))
+        {
+        }
+
+        [JsonSerializable(typeof(AttrClassifiedAnimal))]
+        [JsonSerializable(typeof(BinaryTree))]
+        [JsonSerializable(typeof(ClassifiedAnimalBase))]
+        [JsonSerializable(typeof(ClassifiedAnimalBase[]))]
+        [JsonSerializable(typeof(ClassifiedAnimalDeepHierarchy))]
+        [JsonSerializable(typeof(ClassifiedShape))]
+        [JsonSerializable(typeof(Dictionary<string, ClassifiedAnimalBase>))]
+        [JsonSerializable(typeof(DrawingCanvas))]
+        [JsonSerializable(typeof(FactoryClassifiedAnimal))]
+        [JsonSerializable(typeof(IEnumerable<int>[]))]
+        [JsonSerializable(typeof(IEnumerable<KeyValuePair<int, object>>[]))]
+        [JsonSerializable(typeof(IOpenGenericBase<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_DiamondA<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_DiamondB<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_InterfaceHierarchy<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_InterfaceWrapped<List<string>>))]
+        [JsonSerializable(typeof(IOpenGenericBase_MultiBaseA<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_MultiBaseB<int>))]
+        [JsonSerializable(typeof(IOpenGenericBase_MultiCtor<string>))]
+        [JsonSerializable(typeof(IVarCovBase<VarAnimal>))]
+        [JsonSerializable(typeof(List<ClassifiedAnimalBase>))]
+        [JsonSerializable(typeof(List<PolymorphicClass>))]
+        [JsonSerializable(typeof(List<PolymorphicClassWithCustomTypeDiscriminator>))]
+        [JsonSerializable(typeof(OpenGenericBase<string, int>))]
+        [JsonSerializable(typeof(OpenGenericBase_ArrayArg<int[]>))]
+        [JsonSerializable(typeof(OpenGenericBase_IntDisc<string>))]
+        [JsonSerializable(typeof(OpenGenericBase_KvpArg<KeyValuePair<string, int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_Mixed<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_Multi<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_MultiLevel<List<int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_Partial<string, int>))]
+        [JsonSerializable(typeof(OpenGenericBase_PartiallyConcrete<string, int>))]
+        [JsonSerializable(typeof(OpenGenericBase_Reordered<int, string>))]
+        [JsonSerializable(typeof(OpenGenericBase_StringDisc<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_Tuple<(int, string)>))]
+        [JsonSerializable(typeof(OpenGenericBase_Wrapped<List<string>>))]
+        [JsonSerializable(typeof(Peano))]
+        [JsonSerializable(typeof(PetOwner))]
+        [JsonSerializable(typeof(PolymorphicAbstractClass))]
+        [JsonSerializable(typeof(PolymorphicBaseWithCustomDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass))]
+        [JsonSerializable(typeof(PolymorphicClass[]))]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_NoTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_TypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass_InvalidCustomTypeDiscriminatorPropertyName))]
+        [JsonSerializable(typeof(PolymorphicClass_NoTypeDiscriminators))]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator[]))]
+        [JsonSerializable(typeof(PolymorphicClassWithDerivedCollections))]
+        [JsonSerializable(typeof(PolymorphicClassWithEscapedTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicDictionary))]
+        [JsonSerializable(typeof(PolymorphicInterface))]
+        [JsonSerializable(typeof(PolymorphicInterfaceWithInterfaceDerivedType))]
+        [JsonSerializable(typeof(PolymorphicIntList))]
+        [JsonSerializable(typeof(PolymorphicList))]
+        [JsonSerializable(typeof(PolymorphicTypeWithExtensionDataConflictingPropertyName))]
+        [JsonSerializable(typeof(PolymorphicTypeWithIgnoredConflictingPropertyName))]
+        [JsonSerializable(typeof(TestNode))]
+        [JsonSerializable(typeof(UnsupportedAttrClassifiedAnimal))]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClassWithCustomTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicInterfaceWithInterfaceDerivedType.DerivedClass), TypeInfoPropertyName = "PolymorphicInterfaceWithInterfaceDerivedType_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_CustomConverter_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_CustomConverter_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_CustomConverter_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_InvalidCustomTypeDiscriminatorPropertyName.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_InvalidCustomTypeDiscriminatorPropertyName_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicAbstractClass.Derived), TypeInfoPropertyName = "PolymorphicAbstractClass_Derived")]
+        [JsonSerializable(typeof(PolymorphicClassWithEscapedTypeDiscriminator.Derived), TypeInfoPropertyName = "PolymorphicClassWithEscapedTypeDiscriminator_Derived")]
+        [JsonSerializable(typeof(PolymorphicTypeWithIgnoredConflictingPropertyName.Derived), TypeInfoPropertyName = "PolymorphicTypeWithIgnoredConflictingPropertyName_Derived")]
+        [JsonSerializable(typeof(PolymorphicTypeWithExtensionDataConflictingPropertyName.Derived), TypeInfoPropertyName = "PolymorphicTypeWithExtensionDataConflictingPropertyName_Derived")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass1_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass1_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass1_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass1_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass2_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass2_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClass2_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClass2_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedAbstractClass.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedAbstractClass_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedCollection_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedCollection_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedCollection_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedCollection_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedDictionary_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedDictionary_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedDictionary_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedDictionary_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClassWithConstructor_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClassWithConstructor_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClassWithCustomConverter_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClassWithCustomConverter_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass.DerivedClassWithCustomConverter_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_DerivedClassWithCustomConverter_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicInterface.DerivedClass_NoTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicInterface_DerivedClass_NoTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicInterface.DerivedClass_TypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicInterface_DerivedClass_TypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(List<PolymorphicClassWithConstructor>))]
+        [JsonSerializable(typeof(List<PolymorphicDictionary>))]
+        [JsonSerializable(typeof(List<PolymorphicInterface>))]
+        [JsonSerializable(typeof(List<PolymorphicList>))]
+        [JsonSerializable(typeof(object))]
+        [JsonSerializable(typeof(Customer))]
+        [JsonSerializable(typeof(ObjectWithObjectProperties))]
+        [JsonSerializable(typeof(Person))]
+        [JsonSerializable(typeof(ConstraintBase<List<string>>))]
+        [JsonSerializable(typeof(IVarBivariantBase<VarDog, VarAnimal>))]
+        [JsonSerializable(typeof(IVarContraBase<VarDog>))]
+        [JsonSerializable(typeof(NestedBaseB<NestedOuterB<string>.NestedBoxB<int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_ComplexArg<List<int>>))]
+        [JsonSerializable(typeof(OpenGenericBase_Nested<int>))]
+        [JsonSerializable(typeof(OpenGenericBase_NullableArg<string>))]
+        [JsonSerializable(typeof(OpenGenericBase_StringDisc<string>))]
+        [JsonSerializable(typeof(PolymorphicAbstractClassWithAbstractClassDerivedType))]
+        [JsonSerializable(typeof(PolymorphicClasWithDuplicateTypeDiscriminators))]
+        [JsonSerializable(typeof(PolymorphicClassWithConstructor))]
+        [JsonSerializable(typeof(PolymorphicClassWithDuplicateDerivedTypeRegistrations))]
+        [JsonSerializable(typeof(PolymorphicClassWithNonAssignableDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClassWithObjectDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClassWithStructDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClassWithoutDerivedTypeAttribute))]
+        [JsonSerializable(typeof(PolymorphicClass_WithBaseTypeDiscriminator))]
+        [JsonSerializable(typeof(PolymorphicClass_WithBaseTypeDiscriminator.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_WithBaseTypeDiscriminator_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_WithDerivedPolymorphicClass))]
+        [JsonSerializable(typeof(PolymorphicGenericClass<int>))]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtBase))]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtDerived))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicClass>>))]
+        [JsonSerializable(typeof(GenericPoco<Peano>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClass>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicDictionary>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicInterface>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicList>))]
+        [JsonSerializable(typeof(SimpleTestClassWithObject))]
+        [JsonSerializable(typeof(PolymorphicAbstractClassWithAbstractClassDerivedType.DerivedClass), TypeInfoPropertyName = "PolymorphicAbstractClassWithAbstractClassDerivedType_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClassWithConstructor.DerivedClass), TypeInfoPropertyName = "PolymorphicClassWithConstructor_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClassWithConstructor.DerivedCollection), TypeInfoPropertyName = "PolymorphicClassWithConstructor_DerivedCollection")]
+        [JsonSerializable(typeof(PolymorphicClassWithCustomTypeDiscriminator.DerivedCollection), TypeInfoPropertyName = "PolymorphicClassWithCustomTypeDiscriminator_DerivedCollection")]
+        [JsonSerializable(typeof(PolymorphicClassWithDuplicateDerivedTypeRegistrations.DerivedClass), TypeInfoPropertyName = "PolymorphicClassWithDuplicateDerivedTypeRegistrations_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_NoTypeDiscriminators.DerivedClass2), TypeInfoPropertyName = "PolymorphicClass_NoTypeDiscriminators_DerivedClass2")]
+        [JsonSerializable(typeof(PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass), TypeInfoPropertyName = "PolymorphicClass_WithDerivedPolymorphicClass_DerivedClass")]
+        [JsonSerializable(typeof(PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass.DerivedClass2), TypeInfoPropertyName = "PolymorphicClass_WithDerivedPolymorphicClass_DerivedClass_DerivedClass2")]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtBase.Derived), TypeInfoPropertyName = "PolymorphicTypeWithConflictingPropertyNameAtBase_Derived")]
+        [JsonSerializable(typeof(PolymorphicTypeWithConflictingPropertyNameAtDerived.Derived), TypeInfoPropertyName = "PolymorphicTypeWithConflictingPropertyNameAtDerived_Derived")]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClass>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicDictionary>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicInterface>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicList>))]
+        [JsonSerializable(typeof(List<List<PolymorphicClass>>))]
+        [JsonSerializable(typeof(List<Peano>))]
+        [JsonSerializable(typeof(short[]))]
+        [JsonSerializable(typeof(object[]))]
+        [JsonSerializable(typeof(string[]))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicClassWithConstructor>>))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicDictionary>>))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicInterface>>))]
+        [JsonSerializable(typeof(GenericPoco<List<PolymorphicList>>))]
+        [JsonSerializable(typeof(GenericPoco<object>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClass_WithDerivedPolymorphicClass>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClassWithConstructor>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicClass>>))]
+        [JsonSerializable(typeof(Dictionary<string, string>))]
+        [JsonSerializable(typeof(Dictionary<string, Peano>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClassWithConstructor>))]
+        [JsonSerializable(typeof(HashSet<string>))]
+        [JsonSerializable(typeof(List<List<PolymorphicClassWithConstructor>>))]
+        [JsonSerializable(typeof(List<List<PolymorphicDictionary>>))]
+        [JsonSerializable(typeof(List<List<PolymorphicInterface>>))]
+        [JsonSerializable(typeof(List<List<PolymorphicList>>))]
+        [JsonSerializable(typeof(List<object>))]
+        [JsonSerializable(typeof(List<PolymorphicClass_WithDerivedPolymorphicClass>))]
+        [JsonSerializable(typeof(long[]))]
+        [JsonSerializable(typeof(JsonNode))]
+        [JsonSerializable(typeof(JsonObject))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicClassWithConstructor>>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicDictionary>>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicInterface>>))]
+        [JsonSerializable(typeof(Dictionary<string, List<PolymorphicList>>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClass_WithDerivedPolymorphicClass>))]
+        [JsonSerializable(typeof(Stack<string>))]
+        [JsonSerializable(typeof(System.Collections.IEnumerable))]
+        [JsonSerializable(typeof(ushort[]))]
+        [JsonSerializable(typeof(int[]))]
+        [JsonSerializable(typeof(uint[]))]
+        [JsonSerializable(typeof(ulong[]))]
+        [JsonSerializable(typeof(byte[]))]
+        [JsonSerializable(typeof(sbyte[]))]
+        [JsonSerializable(typeof(char[]))]
+        [JsonSerializable(typeof(bool[]))]
+        [JsonSerializable(typeof(float[]))]
+        [JsonSerializable(typeof(double[]))]
+        [JsonSerializable(typeof(decimal[]))]
+        [JsonSerializable(typeof(System.DateTime[]))]
+        [JsonSerializable(typeof(SampleEnum[]))]
+        [JsonSerializable(typeof(int?[]))]
+        [JsonSerializable(typeof(System.Guid))]
+        [JsonSerializable(typeof(SimpleStruct))]
+        [JsonSerializable(typeof(List<string>))]
+        [JsonSerializable(typeof(Queue<string>))]
+        [JsonSerializable(typeof(LinkedList<string>))]
+        [JsonSerializable(typeof(SortedSet<string>))]
+        [JsonSerializable(typeof(KeyValuePair<string, string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableArray<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableList<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableStack<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableQueue<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableHashSet<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableSortedSet<string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableDictionary<string, string>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableSortedDictionary<string, string>))]
+        [JsonSerializable(typeof(System.Collections.IList))]
+        [JsonSerializable(typeof(System.Collections.ICollection))]
+        [JsonSerializable(typeof(IEnumerable<object>))]
+        [JsonSerializable(typeof(IList<object>))]
+        [JsonSerializable(typeof(ICollection<object>))]
+        [JsonSerializable(typeof(IReadOnlyCollection<object>))]
+        [JsonSerializable(typeof(IReadOnlyList<object>))]
+        [JsonSerializable(typeof(ISet<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableList<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableStack<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableQueue<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.IImmutableSet<object>))]
+        [JsonSerializable(typeof(HashSet<object>))]
+        [JsonSerializable(typeof(Stack<object>))]
+        [JsonSerializable(typeof(Queue<object>))]
+        [JsonSerializable(typeof(LinkedList<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableArray<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableList<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableStack<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableQueue<object>))]
+        [JsonSerializable(typeof(System.Collections.Immutable.ImmutableHashSet<object>))]
+        [JsonSerializable(typeof(GenericPoco<PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass>))]
+        [JsonSerializable(typeof(byte))]
+        [JsonSerializable(typeof(List<PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass>))]
+        [JsonSerializable(typeof(Dictionary<string, PolymorphicClass_WithDerivedPolymorphicClass.DerivedClass>))]
+        [JsonSerializable(typeof(OpenGenericBase_DuplicateDerivedRegistrations<int>))]
+    internal sealed partial class PolymorphicTestsContext_Default : JsonSerializerContext
+        {
+        }
+    }
+
+    public sealed partial class PolymorphicTests_Default_AsyncStream : PolymorphicTests
+    {
+        public PolymorphicTests_Default_AsyncStream()
+            : base(new AsyncStreamSerializerWrapper(PolymorphicTests_Default.PolymorphicTestsContext_Default.Default))
+        {
+        }
+    }
+}
