+++ src/libraries/System.Text.Json/tests/System.Text.Json.Tests/System.Text.Json.Tests.csproj
@@ -98,6 +98,9 @@
     <Compile Include="..\Common\Utf8MemoryStream.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\Utf8MemoryStream.cs" />
     <Compile Include="..\Common\UnmappedMemberHandlingTests.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\UnmappedMemberHandlingTests.cs" />
     <Compile Include="..\Common\UnionTests.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\UnionTests.cs" />
+    <Compile Include="..\Common\PolymorphicTests.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\PolymorphicTests.cs" />
+    <Compile Include="..\Common\PolymorphicTests.CustomTypeHierarchies.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\PolymorphicTests.CustomTypeHierarchies.cs" />
+    <Compile Include="..\Common\PolymorphicTests.TypeClassifier.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\PolymorphicTests.TypeClassifier.cs" />
     <Compile Include="..\Common\TestClasses\TestClasses.ConcurrentCollections.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\TestClasses\TestClasses.ConcurrentCollections.cs" />
     <Compile Include="..\Common\TestClasses\TestClasses.Constructor.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\TestClasses\TestClasses.Constructor.cs" />
     <Compile Include="..\Common\TestClasses\TestClasses.cs" Link="CommonTest\System\Text\Json\Tests\Serialization\TestClasses\TestClasses.cs" />
@@ -226,8 +229,6 @@
     <Compile Include="Serialization\OnSerializeTests.cs" />
     <Compile Include="Serialization\OptionsTests.cs" />
     <Compile Include="Serialization\PolymorphicTests.cs" />
-    <Compile Include="Serialization\PolymorphicTests.CustomTypeHierarchies.cs" />
-    <Compile Include="Serialization\PolymorphicTests.TypeClassifier.cs" />
     <Compile Include="Serialization\StructuralJsonTypeClassifierTests.cs" />
     <Compile Include="Serialization\UnionTests.cs" />
     <Compile Include="Serialization\PropertyNameTests.cs" />
